[trello-27]: Moved from utils to ds dir in pic module
