+++ app/pic/ds/DominantCalculation.py
